protein_drug_discovery/agents.py
import autogen
from typing import Dict, Any, List
from bs4 import BeautifulSoup
import requests
import pymed
from googleapiclient.discovery import build
from protein_drug_discovery.utils import log_wrapper

class SearchAgentWeb(autogen.ConversableAgent):
    def __init__(self, name, **kwargs):
        super().__init__(name, **kwargs)
        self.register_reply(
            [autogen.ConversableAgent, autogen.UserProxyAgent],
            log_wrapper(SearchAgentWeb.search_web),
        )

    @staticmethod
    def search_web(
        recipient: autogen.Agent,
        messages: List[Dict],
        sender: autogen.Agent,
        config: Dict,
    ) -> Dict[str, Any]:

        # For now, we will return a placeholder response.
        # In the future, we will implement the actual web search logic here.
        return {
            "content": "Web search results for protein targets related to the disease.",
            "is_termination_msg": False,
        }

class SearchAgentPubMed(autogen.ConversableAgent):
    def __init__(self, name, **kwargs):
        super().__init__(name, **kwargs)
        self.register_reply(
            [autogen.ConversableAgent, autogen.UserProxyAgent],
            log_wrapper(SearchAgentPubMed.search_pubmed),
        )

    @staticmethod
    def search_pubmed(
        recipient: autogen.Agent,
        messages: List[Dict],
        sender: autogen.Agent,
        config: Dict,
    ) -> Dict[str, Any]:

        # For now, we will return a placeholder response.
        # In the future, we will implement the actual PubMed search logic here.
        return {
            "content": "PubMed search results for protein targets related to the disease.",
            "is_termination_msg": False,
        }

class AnalysisAgent(autogen.ConversableAgent):
    def __init__(self, name, **kwargs):
        super().__init__(name, **kwargs)
        self.register_reply(
            [autogen.ConversableAgent, autogen.UserProxyAgent],
            log_wrapper(AnalysisAgent.analyze_results),
        )

    @staticmethod
    def analyze_results(
        recipient: autogen.Agent,
        messages: List[Dict],
        sender: autogen.Agent,
        config: Dict,
    ) -> Dict[str, Any]:

        # For now, we will return a placeholder response.
        # In the future, we will implement the actual analysis logic here.
        return {
            "content": "Analysis of the search results, providing a list of protein targets.",
            "is_termination_msg": False,
        }

class RankingAgent(autogen.ConversableAgent):
    def __init__(self, name, **kwargs):
        super().__init__(name, **kwargs)
        self.register_reply(
            [autogen.ConversableAgent, autogen.UserProxyAgent],
            log_wrapper(RankingAgent.rank_targets),
        )

    @staticmethod
    def rank_targets(
        recipient: autogen.Agent,
        messages: List[Dict],
        sender: autogen.Agent,
        config: Dict,
    ) -> Dict[str, Any]:

        # For now, we will return a placeholder response.
        # In the future, we will implement the actual ranking logic here.
        return {
            "content": "Ranked list of protein targets with justifications.",
            "is_termination_msg": False,
        }

class CritiqueAgent(autogen.ConversableAgent):
    def __init__(self, name, **kwargs):
        super().__init__(name, **kwargs)
        self.register_reply(
            [autogen.ConversableAgent, autogen.UserProxyAgent],
            log_wrapper(CritiqueAgent.critique_ranking),
        )

    @staticmethod
    def critique_ranking(
        recipient: autogen.Agent,
        messages: List[Dict],
        sender: autogen.Agent,
        config: Dict,
    ) -> Dict[str, Any]:

        # For now, we will return a placeholder response.
        # In the future, we will implement the actual critique logic here.
        return {
            "content": "Critique of the ranked list of protein targets.",
            "is_termination_msg": True,
        }
