protein_drug_discovery/requirements.txt
pyautogen
requests
beautifulsoup4
python-dotenv
pymed
google-api-python-client
